+++ src/game.rs
@@ -2,6 +2,7 @@ use avian2d::prelude::{GravityScale, LinearVelocity};
 use bevy::app::AppExit;
 use bevy::input::common_conditions::*;
 use bevy::prelude::*;
+use bevy::text::{BreakLineOn, Text2dBounds};
 use rand::prelude::*;
 use std::f32::consts::TAU;
 
@@ -10,6 +11,7 @@ use crate::collider::Platform;
 use crate::menu::BLACK_COLOR;
 use crate::spaceship::Player;
 use crate::state::{AppState, GameState};
+use crate::WINDOW_HEIGHT;
 
 pub const FUEL_QUANTITY: f32 = 1000.0;
 
@@ -122,24 +124,39 @@ fn catch_out_of_fuel_event_system(
     mut commands: Commands,
 ) {
     for _event in events_reader.read() {
-        commands.spawn((
-            StateScoped(GameState::Landing),
-            Resettable,
-            TextBundle::from_section(
-                "Out of fuel",
-                TextStyle {
-                    font: assets.font_vt323.clone(),
-                    font_size: 30.0,
+        let box_size = Vec2::new(300.0, 50.0);
+        let box_position = Vec2::new(0.0, WINDOW_HEIGHT / 4.0);
+        commands
+            .spawn((
+                StateScoped(GameState::Landing),
+                Resettable,
+                SpriteBundle {
+                    sprite: Sprite {
+                        color: BLACK_COLOR,
+                        custom_size: Some(Vec2::new(box_size.x, box_size.y)),
+                        ..default()
+                    },
+                    transform: Transform::from_translation(box_position.extend(11.0)),
                     ..default()
                 },
-            )
-            .with_style(Style {
-                position_type: PositionType::Absolute,
-                top: Val::Px(80.0),
-                left: Val::Px(488.0),
-                ..default()
-            }),
-        ));
+            ))
+            .with_children(|builder| {
+                builder.spawn(Text2dBundle {
+                    text: Text::from_section(
+                        "Out of fuel",
+                        TextStyle {
+                            font: assets.font_vt323.clone(),
+                            font_size: 30.0,
+                            color: Color::WHITE,
+                            ..default()
+                        },
+                    )
+                    .with_justify(JustifyText::Left),
+                    text_2d_bounds: Text2dBounds { size: box_size },
+                    transform: Transform::from_translation(Vec3::Z),
+                    ..default()
+                });
+            });
     }
 }